First working version of simulate()
The thread function was moved outside of the SpinLattice class to
a separate function. Now, the lattice is passed by a pointer to the
thread function and the main thread pinkie swears not to modify it.
A better solution would be to move the lattice into simulate(), but
I couldn't make it work.